+++ packages/angular-workspace/example-client-app/src/app/app.module.ts
@@ -34,12 +34,9 @@ import { NimbleTableColumnMenuButtonModule } from '@ni/nimble-angular/table-colu
 import { NimbleRichTextViewerModule } from '@ni/nimble-angular/rich-text/viewer';
 import { NimbleRichTextEditorModule } from '@ni/nimble-angular/rich-text/editor';
 import { NimbleRichTextMentionUsersModule } from '@ni/nimble-angular/rich-text-mention/users';
-import { OkFvAccordionItemModule } from 'ok-angular/fv-accordion-item/fv-accordion-item.module';
 import { NimbleStepperModule } from '@ni/nimble-angular/stepper';
 import { NimbleStepModule } from '@ni/nimble-angular/step';
 import { NimbleAnchorStepModule } from '@ni/nimble-angular/anchor-step';
-import { OkButtonModule } from 'ok-angular/button/ok-button.module';
-import { OkFvSearchInputModule } from 'ok-angular/fv-search-input/fv-search-input.module';
 import { SprightChatConversationModule } from '@ni/spright-angular/chat/conversation';
 import { SprightChatInputModule } from '@ni/spright-angular/chat/input';
 import { SprightIconWorkItemCalendarWeekDirective } from '@ni/spright-angular/icons/work-item-calendar-week';
@@ -49,7 +46,7 @@ import { SprightChatMessageSystemModule } from '@ni/spright-angular/chat/message
 import { SprightRectangleModule } from '@ni/spright-angular/rectangle';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
-import { SubContainerComponent } from './customapp/sub-container.component';
+import { SubContainerModule } from './customapp/sub-container/sub-container.module';
 import { OverviewSectionComponent } from './customapp/overview-section.component';
 import { AnchorSectionComponent } from './customapp/anchor-section.component';
 import { BannerSectionComponent } from './customapp/banner-section.component';
@@ -87,15 +84,13 @@ import { TreeViewSectionComponent } from './customapp/tree-view-section.componen
 import { ChatConversationSectionComponent } from './customapp/chat-conversation-section.component';
 import { IconsSectionComponent } from './customapp/icons-section.component';
 import { RectangleSectionComponent } from './customapp/rectangle-section.component';
-import { OkButtonSectionComponent } from './customapp/ok-button-section.component';
-import { OkAccordionItemSectionComponent } from './customapp/ok-accordion-item-section.component';
-import { OkFvSearchInputSectionComponent } from './customapp/ok-fv-search-input-section.component';
+import { ExSectionModule } from './customapp/ex/ex-section.module';
+import { FvSectionModule } from './customapp/fv/fv-section.module';
 
 @NgModule({
     declarations: [
         AppComponent,
         CustomAppComponent,
-        SubContainerComponent,
         OverviewSectionComponent,
         AnchorSectionComponent,
         BannerSectionComponent,
@@ -132,10 +127,7 @@ import { OkFvSearchInputSectionComponent } from './customapp/ok-fv-search-input-
         TreeViewSectionComponent,
         ChatConversationSectionComponent,
         IconsSectionComponent,
-        RectangleSectionComponent,
-        OkButtonSectionComponent,
-        OkAccordionItemSectionComponent,
-        OkFvSearchInputSectionComponent
+        RectangleSectionComponent
     ],
     imports: [
         BrowserModule,
@@ -207,9 +199,8 @@ import { OkFvSearchInputSectionComponent } from './customapp/ok-fv-search-input-
         NimbleMappingEmptyModule,
         NimbleIconPencilToRectangleModule,
         NimbleIconMessagesSparkleModule,
-        OkFvAccordionItemModule,
-        OkButtonModule,
-        OkFvSearchInputModule,
+        FvSectionModule,
+        ExSectionModule,
         SprightChatConversationModule,
         SprightChatInputModule,
         SprightChatMessageInboundModule,
@@ -220,6 +211,7 @@ import { OkFvSearchInputSectionComponent } from './customapp/ok-fv-search-input-
         NimbleStepperModule,
         NimbleStepModule,
         NimbleAnchorStepModule,
+        SubContainerModule,
         RouterModule.forRoot(
             [
                 { path: '', redirectTo: '/customapp', pathMatch: 'full' },