+++ packages/ok-components/src/all-components.ts
@@ -6,7 +6,6 @@
 
 import '@ni/spright-components/dist/esm/all-components';
 
-import './fv-accordion-item';
-import './button';
-import './icon-dynamic';
-import './fv-search-input';
+import './fv/all-fv';
+import './ex/all-ex';
+import './ts/all-ts';