+++ packages/react-workspace/react-client-app/src/components/App.tsx
@@ -41,9 +41,8 @@ import { TooltipSection } from './TooltipSection';
 import { TreeViewSection } from './TreeViewSection';
 import { ChatConversationSection } from './ChatConversationSection';
 import { RectangleSection } from './RectangleSection';
-import { OkAccordionItemSection } from './OkAccordionItemSection';
-import { OkButtonSection } from './OkButtonSection';
-import { OkFvSearchInputSection } from './OkFvSearchInputSection';
+import { ExSection } from './Ex/ExSection';
+import { FvSection } from './Fv/FvSection';
 
 export function App(): React.JSX.Element {
     const prefersColorSchemeDarkMediaQuery: MediaQueryList = window.matchMedia(
@@ -111,9 +110,8 @@ export function App(): React.JSX.Element {
                                 <TreeViewSection />
                                 <ChatConversationSection />
                                 <RectangleSection />
-                                <OkAccordionItemSection />
-                                <OkButtonSection />
-                                <OkFvSearchInputSection />
+                                <FvSection />
+                                <ExSection />
                             </div>
                         </div>
                     </div>