+++ packages/react-workspace/react-client-app/src/components/Fv/FvSection.tsx
@@ -1,0 +1,11 @@
+import { FvAccordionItemSection } from './FvAccordionItemSection';
+import { OkFvSearchInputSection } from '../OkFvSearchInputSection';
+
+export function FvSection(): React.JSX.Element {
+    return (
+        <>
+            <FvAccordionItemSection></FvAccordionItemSection>
+            <OkFvSearchInputSection></OkFvSearchInputSection>
+        </>
+    );
+}