+++ packages/spright-components/src/chat/input/testing/chat-input.pageobject.ts
@@ -86,6 +86,22 @@ export class ChatInputPageObject {
         return this.element.textArea!.value;
     }
 
+    public hasFooterActionsSlot(): boolean {
+        return this.getFooterActionsSlot() !== null;
+    }
+
+    public getFooterActionsSlotElementCount(): number {
+        return this.getFooterActionsSlot()?.assignedElements().length ?? 0;
+    }
+
+    public hasAttachmentSlot(): boolean {
+        return this.getAttachmentsSlot() !== null;
+    }
+
+    public getAttachmentsSlotElementCount(): number {
+        return this.getAttachmentsSlot()?.assignedElements().length ?? 0;
+    }
+
     public setText(text: string): void {
         this.element.textArea!.focus();
         this.element.textArea!.value = text;
@@ -106,6 +122,16 @@ export class ChatInputPageObject {
     private getActionButton(): Button {
         const actionButton = this.element.shadowRoot!.querySelector<Button>('.action-button')!;
         return actionButton;
+    }
+
+    private getAttachmentsSlot(): HTMLSlotElement | null {
+        const attachmentsSlot = this.element.shadowRoot!.querySelector<HTMLSlotElement>('slot[name="attachments"]');
+        return attachmentsSlot;
+    }
+
+    private getFooterActionsSlot(): HTMLSlotElement | null {
+        const footerActionsSlot = this.element.shadowRoot!.querySelector<HTMLSlotElement>('slot[name="footer-actions"]');
+        return footerActionsSlot;
     }
 
     private async sendEnterKeyEvents(shiftKey: boolean): Promise<void> {