+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @ni/spright-components
 
-<!-- This log was last generated on Mon, 20 Apr 2026 16:30:31 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 21 Apr 2026 19:47:00 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 6.17.0
+
+Tue, 21 Apr 2026 19:47:00 GMT
+
+### Minor changes
+
+- add footer-actions and attachments slots ([ni/nimble@ead5be1](https://github.com/ni/nimble/commit/ead5be103d14e946bff31bfa222c168e966fca88))
+
+## 6.16.0
+
+Mon, 20 Apr 2026 22:42:19 GMT
+
+### Minor changes
+
+- Add welcome message for spright chat conversation ([ni/nimble@f7e4581](https://github.com/ni/nimble/commit/f7e4581eb131183f6ebccd9f6337b1f54eb9be8f))
 
 ## 6.15.9